Add .gitreview to make git-review work w/ no conf
With defaultremote set, git-review requires no configuration beyond
cloning. It will *not* need a remote called "gerrit", and it will not
automatically add a new remote the first time git-review is invoked.
This is the effect I desired to get from this changeset. We also provide
the host and project settings to be specific.

Change-Id: Id19f3482fff4e1cde469b92e11f54eab7b704633